list separated in directory
